Bump depend/bitcoin to c84c752506d750e0d9c87a58ada9cbc19fc7a83b

+++ depend/bitcoin/src/rpc/fees.cpp
@@ -1,5 +1,5 @@
 // Copyright (c) 2010 Satoshi Nakamoto
-// Copyright (c) 2009-2022 The Bitcoin Core developers
+// Copyright (c) 2009-present The Bitcoin Core developers
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
@@ -195,14 +195,14 @@ static RPCHelpMan estimaterawfee()
                 if (feeRate != CFeeRate(0)) {
                     horizon_result.pushKV("feerate", ValueFromAmount(feeRate.GetFeePerK()));
                     horizon_result.pushKV("decay", buckets.decay);
-                    horizon_result.pushKV("scale", (int)buckets.scale);
+                    horizon_result.pushKV("scale", buckets.scale);
                     horizon_result.pushKV("pass", std::move(passbucket));
                     // buckets.fail.start == -1 indicates that all buckets passed, there is no fail bucket to output
                     if (buckets.fail.start != -1) horizon_result.pushKV("fail", std::move(failbucket));
                 } else {
                     // Output only information that is still meaningful in the event of error
                     horizon_result.pushKV("decay", buckets.decay);
-                    horizon_result.pushKV("scale", (int)buckets.scale);
+                    horizon_result.pushKV("scale", buckets.scale);
                     horizon_result.pushKV("fail", std::move(failbucket));
                     errors.push_back("Insufficient data or no feerate found which meets threshold");
                     horizon_result.pushKV("errors", std::move(errors));
